Rename AlwaysHealthy to SilentHealth and remove Health.Adapter
