+++ community/common/src/main/java/org/neo4j/kernel/SilentHealth.java
@@ -19,6 +19,35 @@
  */
 package org.neo4j.kernel;
 
-public class AlwaysHealthy extends Health.Adapter
+public class SilentHealth implements Health
 {
+    @Override
+    public <EXCEPTION extends Throwable> void assertHealthy( Class<EXCEPTION> panicDisguise ) throws EXCEPTION
+    {
+        // no-op
+    }
+
+    @Override
+    public void panic( Throwable cause )
+    {
+        // no-op
+    }
+
+    @Override
+    public boolean isHealthy()
+    {
+        return true;
+    }
+
+    @Override
+    public void healed()
+    {
+        // no-op
+    }
+
+    @Override
+    public Throwable cause()
+    {
+        return null;
+    }
 }
