+++ community/index/src/test/java/org/neo4j/index/internal/gbptree/GBPTreeIT.java
@@ -36,7 +36,7 @@ import java.util.concurrent.Executors;
 import org.neo4j.cursor.RawCursor;
 import org.neo4j.io.pagecache.IOLimiter;
 import org.neo4j.io.pagecache.PageCache;
-import org.neo4j.kernel.AlwaysHealthy;
+import org.neo4j.kernel.SilentHealth;
 import org.neo4j.logging.NullLog;
 import org.neo4j.test.rule.PageCacheRule;
 import org.neo4j.test.rule.RandomRule;
@@ -78,7 +78,7 @@ public class GBPTreeIT
     {
         pageCache = pageCacheRule.getPageCache( fs.get(), config().withPageSize( pageSize ).withAccessChecks( true ) );
         return index = new GBPTree<>( pageCache, directory.file( "index" ),
-                layout, 0/*use whatever page cache says*/, monitor, NO_HEADER, new AlwaysHealthy(),
+                layout, 0/*use whatever page cache says*/, monitor, NO_HEADER, new SilentHealth(),
                 NullLog.getInstance() );
     }
